Refactor withdrawal activity boundaries
