+++ src/lib/bridge/withdrawalActivityValues.js
@@ -1,0 +1,94 @@
+export const MAX_ACTIVITY_ITEMS = 12;
+
+const EVM_ADDRESS_PATTERN = /^0x[0-9a-fA-F]{40}$/;
+const TX_HASH_PATTERN = /^0x[0-9a-fA-F]{64}$/;
+
+/**
+ * @typedef {object} WithdrawalActivityItem
+ * @property {string | null} account
+ * @property {string | null} amountWei
+ * @property {string | null} blockNumber
+ * @property {string | null} challengePeriodEnd
+ * @property {number} chainId
+ * @property {number} createdAt
+ * @property {string | null} explorerStatus
+ * @property {string | null} l1TransactionHash
+ * @property {"explorer" | "local"} source
+ * @property {string} stage
+ * @property {string | null} timestamp
+ * @property {`0x${string}`} transactionHash
+ */
+
+/**
+ * @param {unknown} value
+ */
+export function errorMessage(value) {
+  return value instanceof Error ? value.message : String(value);
+}
+
+/**
+ * @param {unknown} value
+ */
+function addressHash(value) {
+  if (typeof value === "string") {
+    return value;
+  }
+
+  if (value && typeof value === "object" && "hash" in value) {
+    return typeof value.hash === "string" ? value.hash : null;
+  }
+
+  return null;
+}
+
+/**
+ * @param {unknown} value
+ */
+export function normalizedAddress(value) {
+  const hash = addressHash(value);
+
+  return hash && EVM_ADDRESS_PATTERN.test(hash) ? hash.toLowerCase() : null;
+}
+
+/**
+ * @param {unknown} value
+ * @returns {`0x${string}` | null}
+ */
+export function normalizedTxHash(value) {
+  return typeof value === "string" && TX_HASH_PATTERN.test(value)
+    ? /** @type {`0x${string}`} */ (value.toLowerCase())
+    : null;
+}
+
+/**
+ * @param {unknown} value
+ */
+export function normalizedTimestamp(value) {
+  if (typeof value !== "string" && typeof value !== "number") {
+    return null;
+  }
+
+  const date = new Date(value);
+
+  return Number.isNaN(date.getTime()) ? null : date.toISOString();
+}
+
+/**
+ * @param {unknown} value
+ */
+export function stringValue(value) {
+  return typeof value === "string" || typeof value === "number"
+    ? String(value)
+    : null;
+}
+
+/**
+ * @param {unknown} value
+ */
+export function normalizedAmountWei(value) {
+  if (typeof value === "bigint") {
+    return value >= 0n ? value.toString() : null;
+  }
+
+  return typeof value === "string" && /^\d+$/u.test(value) ? value : null;
+}
